refactor: reorganize component imports and update paths for better structure
