+++ src/app/layout.tsx
@@ -3,8 +3,8 @@
 import { Inter } from "next/font/google";
 import "./globals.css";
 import dynamic from "next/dynamic";
-import NextAuthProvider from "../components/next-auth-provider";
-import MiniKitProvider from "../components/minikit-provider";
+import NextAuthProvider from "../providers/next-auth-provider";
+import MiniKitProvider from "../providers/minikit-provider";
 
 const inter = Inter({ subsets: ["latin"] });
